// src/models/short-knowledge-base.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger Data Provider
 * Документация API Провайдера Данных
 *
 * OpenAPI spec version: 1.0.0
 * 
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

 /**
 * Knowledge Base
 *
 * @export
 * @interface ShortKnowledgeBase
 */
export interface ShortKnowledgeBase {

    /**
     * Knowledge Base identifier
     *
     * @type {string}
     * @memberof ShortKnowledgeBase
     * @example d62fad7a-2d00-456b-b711-67dfbfd4f07f
     */
    id: string;

    /**
     * Knowledge Base name
     *
     * @type {string}
     * @memberof ShortKnowledgeBase
     */
    name: string;

    /**
     * Knowledge Base description
     *
     * @type {string}
     * @memberof ShortKnowledgeBase
     */
    description: string;

    /**
     * Type of knowledge base
     *
     * @type {string}
     * @memberof ShortKnowledgeBase
     * @example FILE
     */
    type: ShortKnowledgeBaseTypeEnum;

    /**
     * Icon url
     *
     * @type {string}
     * @memberof ShortKnowledgeBase
     */
    iconUrl?: string;

    /**
     * Number of maximum users for knowledge base
     *
     * @type {number}
     * @memberof ShortKnowledgeBase
     * @example 10
     */
    maxUsers?: number;

    /**
     * Number of current users for knowledge base
     *
     * @type {number}
     * @memberof ShortKnowledgeBase
     * @example 5
     */
    currentUsers?: number;
}

/**
 * @export
 * @enum {string}
 */
export enum ShortKnowledgeBaseTypeEnum {
    FILE = 'FILE',
    HSE = 'HSE',
    CREDENTIALS = 'CREDENTIALS'
}