// src/models/knowledge-base.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Swagger Data Provider
 * Документация API Провайдера Данных
 *
 * OpenAPI spec version: 1.0.0
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

import { KnowledgeBaseData } from "./knowledge-base-data";
import { KnowledgeBaseSession } from "./knowledge-base-session";
/**
 * Knowledge Base
 *
 * @export
 * @interface KnowledgeBase
 */
export interface KnowledgeBase {
  /**
   * Knowledge Base identifier
   *
   * @type {string}
   * @memberof KnowledgeBase
   * @example d62fad7a-2d00-456b-b711-67dfbfd4f07f
   */
  id: string;

  /**
   * Knowledge Base name
   *
   * @type {string}
   * @memberof KnowledgeBase
   */
  name: string;

  /**
   * Knowledge Base description
   *
   * @type {string}
   * @memberof KnowledgeBase
   */
  description: string;

  /**
   * Type of knowledge base
   *
   * @type {string}
   * @memberof KnowledgeBase
   * @example FILE
   */
  type: KnowledgeBaseTypeEnum;

  /**
   * Icon url
   *
   * @type {string}
   * @memberof KnowledgeBase
   */
  iconUrl?: string;

  /**
   * Redirect url
   *
   * @type {string}
   * @memberof KnowledgeBase
   */
  redirectUrl?: string;

  /**
   * Instruction for use
   *
   * @type {string}
   * @memberof KnowledgeBase
   */
  instruction?: string;

  /**
   * Max session time in millis
   *
   * @type {number}
   * @memberof KnowledgeBase
   */
  sessionTimeMillis?: number;

  /**
   * Number of maximum users for knowledge base
   *
   * @type {number}
   * @memberof KnowledgeBase
   * @example 10
   */
  maxUsers?: number;

  /**
   * Number of current users for knowledge base
   *
   * @type {number}
   * @memberof KnowledgeBase
   * @example 5
   */
  currentUsers?: number;

  /**
   * @type {KnowledgeBaseData}
   * @memberof KnowledgeBase
   */
  data?: KnowledgeBaseData;

  /**
   * @type {KnowledgeBaseSession}
   * @memberof KnowledgeBase
   */
  activeSession?: KnowledgeBaseSession;
}

/**
 * @export
 * @enum {string}
 */
export enum KnowledgeBaseTypeEnum {
  FILE = "FILE",
  INSTRUCTION = "INSTRUCTION",
  CREDENTIALS = "CREDENTIALS",
}